Migrate test stripping to pytest
